Remove VS project files so that they don't have to be maintained
These complicate extern upgrades and tweaks to the code structure, e.g.
when we add shaders we need to replicate them in both cases. CMake works
just fine on Windows without extra setup thanks to Visual Studio
supporting it.

Also update README.

+++ src/niagara.cpp
@@ -1169,7 +1169,7 @@ int main(int argc, const char** argv)
 		cullData.frustum[1] = frustumX.z;
 		cullData.frustum[2] = frustumY.y;
 		cullData.frustum[3] = frustumY.z;
-		cullData.drawCount = draws.size();
+		cullData.drawCount = uint32_t(draws.size());
 		cullData.cullingEnabled = cullingEnabled;
 		cullData.lodEnabled = lodEnabled;
 		cullData.occlusionEnabled = occlusionEnabled;
